requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
qrcode==7.4.2
Pillow==12.1.0
python-dotenv==1.0.0
Flask==3.0.0
Flask-Cors==4.0.0
mysql-connector-python==8.2.0
qrcode==7.4.2
Pillow==12.1.0
python-dotenv==1.0.0